requirements.txt
-i https://pypi.org/simple
certifi==2019.6.16
django-minio-storage==0.2.2
django==2.2.4
djangorestframework==3.10.2
future==0.17.1
minio==4.0.20
pillow==6.1.0
postgres==2.2.2
psycopg2-binary==2.8.3
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
